pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = ["molodost-vnutri <90547216+molodost-vnutri@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
SQLAlchemy = "^2.0.31"
fastapi = "^0.111.1"
PyJWT = "^2.8.0"
passlib = "^1.7.4"
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
